use framework bundle translation resource files already generated
Instead of using the old way of manually fetching folder by folder using the finder.
This way we maintain the same behaviour of the Translator support in symfony, because they all detect the same files.
This hotfix also fixes if someone wants to change the translations files folder path in the framework.translator.paths configuration inside the framework bundle.
Supports also old versions of Symfony using a Finder to fetch the files.