Fix Travis configuration for different distributions
HHVM is only available on trusty now, not on precise. And precise does not have PHP 5.3.

So this forces the distribution explicitly for both of them.
This fixes the HHVM job today, and ensures that the PHP 5.3 job keeps working when Travis migrates this repository to using trusty by default (they are currently deploying this upgrade little by little)?