fix: ensure fallback locale is always loaded
In the current version, when a fallback locale is defined, it isn't loaded automatically (as explained in #307). 

That means you still have to specifically request that the fallback be loaded.

This commit updates the `getLocales()` method to ensure the fallback translations are always loaded when a fallback locale is defined.

Here's a comparison in the case of a project with the `fr_CA` as the default locale and `fr` as the fallback locale (with only one message, in order to keep things simple).

**Before:**
```js
(function (t) {
t.fallback = 'fr';
t.defaultDomain = 'messages';
// fr_CA
})(Translator);
```

**After:**
```js
(function (t) {
t.fallback = 'fr';
t.defaultDomain = 'messages';
// fr_CA
// fr
t.add("generic.save", "Enregistrer", "messages", "fr");
})(Translator);
```

Closes #307