Replace the JS test runner (#367)
* Upgrade the JS testsuite to use QUnit 2

This uses npm to install the testsuite dependencies instead of vendoring
them in the project to make the setup easier to maintain.
This also removes the file loading the ES5-shim as all maintained
browsers support ES5 natively anyway.

* Modernize the UMD structure for the translator

* Replace the test runner for JS tests

PhantomJS is not maintained anymore and the setup was broken. The new
setup uses `@web/test-runner` to run the QUnit tests in an actual
browser.